+++ assetconnection/common/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/common/ElementInfo.java
@@ -36,7 +36,7 @@ public class ElementInfo {
     }
 
 
-    public TypeInfo<?> getTypeInfo() {
+    public TypeInfo getTypeInfo() {
         return typeInfo;
     }
